+++ source/animation.h
@@ -10,7 +10,7 @@ void shutdown_animations();
 
 void remove_from_animations(entity_t *);
 
-void process_animations(uint64_t);
+void process_animations(uint16_t);
 
 void animation_rollup(entity_t *);
 void animation_rolldown(entity_t *);
