+++ source/entity.h
@@ -8,7 +8,7 @@
 int init_entities();
 void shutdown_entities();
 
-void entities_logic(rectangle_t const *, uint64_t);
+void entities_logic(rectangle_t const *, uint16_t);
 void sprites_update(rectangle_t const *);
 
 template_t * template_new(const char *, uint16_t, uint16_t, uint8_t, uint8_t, uint8_t, texture_t const *, const char *);
